Fix gcc string truncation warning for strncpy
If the length of the source string is equal to or greater than this size the result of the copy
will not be NUL-terminated. Therefore, the call is also diagnosed. To avoid the warning,
specify sizeof buf - 1 as the bound and set the last element of the buffer to NUL.

See:
https://gcc.gnu.org/onlinedocs/gcc-8.1.0/gcc/Warning-Options.html#index-Wstringop-truncation

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>