tests: Accept all causes for UnknownHostException
In some build environments, networking is restricted or disabled, which
can cause the UnknownHostException to be thrown with varying causes and
error messages. This commit filters out the error message from the test
result and also removes it from the reference files that we use to
evaluate the test. This way, we only check that the exception is thrown
as expected and the test doesn't fail when it's thrown for an unexpected
reason, e.g. "Name or service not known" vs. "Temporary failure in name
resolution".

See also bdacba19a48199434137961a555fe5755149792b.